+++ src/Scripts/AudioLogic.lua
@@ -10,7 +10,7 @@ modutil.mod.Path.Wrap("StartRoomAmbience", function(base, currentRun, currentRoo
 end)
 
 modutil.mod.Path.Wrap("AudioStateInit", function(base, triggerArgs)
-	local currentRoom = nil
+	local currentRoom = {}
 	if game.CurrentRun ~= nil and game.CurrentRun.CurrentRoom ~= nil then
 		currentRoom = game.CurrentRun.CurrentRoom
 	end
@@ -39,4 +39,15 @@ modutil.mod.Path.Wrap("AudioStateInit", function(base, triggerArgs)
 	end
 
 	return base(triggerArgs)
+end)
+
+modutil.mod.Path.Wrap("ResumeMusic", function(base, args)
+	if game.CurrentRun and game.CurrentRun.ModsNikkelMHadesBiomesIsModdedRun then
+		-- Sometimes after challenge encounters, the next room has doubled music, which is most likely caused by a double ResumeMusic at inopportune times
+		if game.AudioState.MusicId == nil or not game.AudioState.MusicPaused then
+			return
+		end
+	end
+
+	return base(args)
 end)