+++ src/Scripts/RoomData.lua
@@ -16,17 +16,84 @@ mod.ApplyModificationsAndInheritRoomData(baseRoomData, baseRoomModifications, ba
 -- #region CHALLENGE
 local challengeRoomData = {
 	RoomChallenge01 = mod.RoomData.Base.RoomChallenge01,
+	RoomChallenge02 = mod.RoomData.Base.RoomChallenge02,
+	RoomChallenge03 = mod.RoomData.Base.RoomChallenge03,
+	RoomChallenge04 = mod.RoomData.Base.RoomChallenge04,
 }
 
 local challengeRoomReplacements = {}
 local challengeRoomModifications = {
 	RoomChallenge01 = {
-		RewardPreviewIcon = "RoomRewardSubIcon_Miniboss",
+		RoomSetName = "Challenge",
+
+		LoadModdedPackages = { "RoomManagerModsNikkelMHadesBiomes", "NikkelM-HadesBiomesFxModded", "NikkelM-HadesBiomesGUIModded", "ModsNikkelMHadesBiomesGUIOriginal", },
+		LoadModdedAudioBanks = { "EnemiesModsNikkelMHadesBiomes", "SoundsModsNikkelMHadesBiomes", },
+		LoadCustomModdedAudioBanks = { "ModsNikkelMHadesBiomesMusicModded", },
+
+		SaveProfileLocationText = "ModsNikkelMHadesBiomesLocation_Hades_Challenge",
+
+		LocationAnimName = "ModsNikkelMHadesBiomesInfoBannerErebusIn",
+		LocationAnimOutName = "ModsNikkelMHadesBiomesInfoBannerErebusOut",
+
+		NextRoomResumeMusic = true,
+		EntranceVfx = "SecretDoorDiveFx",
+		ExitFunctionName = _PLUGIN.guid .. "." .. "ModsNikkelMHadesBiomesShrineGateExitPresentation",
+		ExitAnimation = "MelinoeSpellFire",
+		ExitVfx = "SecretDoorDiveFx",
+
+		IneligibleRewards = { "Devotion", "WeaponUpgrade", "SpellDrop", "HermesUpgrade" },
+		-- So it's caught by the custom logic for upgraded consumable rewards
+		UseOptionalOverrides = true,
+		-- Need to define this to prevent nil error
+		OptionalOverrides = {},
+		-- We directly change the consumables to their upgraded versions when choosing the door reward using the map
+		RewardConsumableOverrides = mod.NilValue,
+		RewardConsumableOverrideMap = {
+			RoomMoneyDrop = "RoomMoneyBigDrop",
+			MaxHealthDrop = "MaxHealthDropBig",
+			MaxManaDrop = "MaxManaDropBig",
+			StackUpgrade = "StackUpgradeBig",
+			TalentDrop = "TalentBigDrop",
+		},
+
+		InspectPoints = {
+			[515865] = {
+				InteractTextLineSets = {
+					CharonFirstInspect = mod.NilValue,
+					ModsNikkelMHadesBiomes_Challenge01FirstInspect = {
+						-- Is this where Father's goons assemble to come after me?
+						EndCue = "/VO/ZagreusField_1472",
+						EndWait = 0.3,
+						{
+							Cue = "/VO/Storyteller_0220",
+							Text =
+							"{#DialogueItalicFormat}Concealed in the dark recesses of the Underworld is the land of Erebus, wherein the dead await eternal sentencing; they long for the attention and good graces of their master, the Lord Hades."
+						},
+					},
+				},
+			},
+		},
 	},
 }
 
 mod.ApplyModificationsAndInheritRoomData(challengeRoomData, challengeRoomModifications, challengeRoomReplacements,
 	"Challenge")
+-- #endregion
+
+-- #region CHARON
+local charonRoomData = {
+	CharonFight01 = mod.RoomData.Base.CharonFight01,
+}
+
+local charonRoomReplacements = {}
+local charonRoomModifications = {
+	CharonFight01 = {
+		-- The encounter has BiomeName = "Challenge" - should this be Challenge or separate/Charon? How does it work with the MapGroups.sjson?
+		-- RoomSetName = "Charon",
+	},
+}
+
+-- mod.ApplyModificationsAndInheritRoomData(charonRoomData, charonRoomModifications, charonRoomReplacements, "Charon")
 -- #endregion
 
 -- #region SURFACE